+++ src/moderation/moderation.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { AuthModule } from '../auth/auth.module';
 import { OffersModule } from '../offers/offers.module';
+import { AdminCommentsController } from './admin-comments.controller';
 import { AdminOffersController } from './admin-offers.controller';
 import { AdminReportsController } from './admin-reports.controller';
 import { AdminUsersController } from './admin-users.controller';
@@ -13,6 +14,7 @@ import { ModerationService } from './moderation.service';
     AdminOffersController,
     AdminUsersController,
     AdminReportsController,
+    AdminCommentsController,
   ],
   providers: [ModerationService],
 })